# [IC-705] Optimize socket and audio handling