--- a/firmware/main/network/icom/AudioState.cpp
+++ b/firmware/main/network/icom/AudioState.cpp
@@ -23,6 +23,196 @@
 #include "AudioState.h"
 #include "IcomStateMachine.h"
 
+// The below values are calculated values for each possible audio multiplier
+// in terms of Q5.11 fixed point. The formula for calculating these is:
+//
+//     e^(volInDb/20.0 * ln(10.0))
+//
+// And the equivalent fixed point values are calculated by using the calculator
+// at https://chummersone.github.io/qformat.html.
+//
+// (Note: ezDV operates in the range of -63.5dB to +24dB amplification.)
+static short FIXED_POINT_AMPLIFICATION_FACTORS[] = {
+    1,
+    1,
+    2,
+    2,
+    2,
+    2,
+    2,
+    2,
+    2,
+    2,
+    2,
+    3,
+    3,
+    3,
+    3,
+    3,
+    3,
+    4,
+    4,
+    4,
+    4,
+    5,
+    5,
+    5,
+    5,
+    6,
+    6,
+    6,
+    7,
+    7,
+    8,
+    8,
+    9,
+    9,
+    10,
+    10,
+    11,
+    12,
+    12,
+    13,
+    14,
+    14,
+    15,
+    16,
+    17,
+    18,
+    19,
+    20,
+    22,
+    23,
+    24,
+    26,
+    27,
+    29,
+    31,
+    32,
+    34,
+    36,
+    39,
+    41,
+    43,
+    46,
+    49,
+    51,
+    54,
+    58,
+    61,
+    65,
+    69,
+    73,
+    77,
+    82,
+    86,
+    91,
+    97,
+    103,
+    109,
+    115,
+    122,
+    129,
+    137,
+    145,
+    154,
+    163,
+    172,
+    183,
+    193,
+    205,
+    217,
+    230,
+    243,
+    258,
+    273,
+    289,
+    306,
+    325,
+    344,
+    364,
+    386,
+    409,
+    433,
+    458,
+    486,
+    514,
+    545,
+    577,
+    611,
+    648,
+    686,
+    727,
+    770,
+    815,
+    864,
+    915,
+    969,
+    1026,
+    1087,
+    1152,
+    1220,
+    1292,
+    1369,
+    1450,
+    1536,
+    1627,
+    1723,
+    1825,
+    1933,
+    2048,
+    2169,
+    2298,
+    2434,
+    2578,
+    2731,
+    2893,
+    3064,
+    3246,
+    3438,
+    3642,
+    3858,
+    4086,
+    4328,
+    4585,
+    4857,
+    5144,
+    5449,
+    5772,
+    6114,
+    6476,
+    6860,
+    7267,
+    7697,
+    8153,
+    8636,
+    9148,
+    9690,
+    10264,
+    10873,
+    11517,
+    12199,
+    12922,
+    13688,
+    14499,
+    15358,
+    16268,
+    17232,
+    18253,
+    19334,
+    20480,
+    21694,
+    22979,
+    24341,
+    25783,
+    27311,
+    28929,
+    30643,
+};
+
+#define MIN_AMPLIFICATION_DB (-127) /* -63.5dB minimum amplification by TLV320 */
+#define UNITY_AMPLIFICATION_VAL (2048) /* 1.0 */
+
 namespace ezdv
 {
 
@@ -44,7 +234,7 @@ AudioState::AudioState(IcomStateMachine* parent)
 
     for (int index = 0; index < 160; index++)
     {
-        audioMultiplier_[index] = 1;
+        audioMultiplier_[index] = UNITY_AMPLIFICATION_VAL;
     }
 }
 
@@ -122,26 +312,18 @@ void AudioState::onAudioOutTimer_(DVTimer*)
     // Get input audio and write to socket
     uint16_t samplesToRead = 160; // 320 bytes
     short tempAudioOut[samplesToRead];
-    float tempAudioInFloat[samplesToRead];
-    float tempAudioOutFloat[samplesToRead];
     //memset(tempAudioOut, 0, samplesToRead * sizeof(short));
 
     if (codec2_fifo_used(inputFifo) >= samplesToRead)
     {
         codec2_fifo_read(inputFifo, tempAudioOut, samplesToRead);
 
-        for (int index = 0; index < samplesToRead; index++)
-        {
-            tempAudioInFloat[index] = tempAudioOut[index];
-        }
-
-        // Adjust output based on configured volume
-        dsps_mul_f32(tempAudioInFloat, audioMultiplier_, tempAudioOutFloat, samplesToRead, 1, 1, 1);
-
-        for (int index = 0; index < samplesToRead; index++)
-        {
-            tempAudioOut[index] = tempAudioOutFloat[index];
-        }
+        // Adjust output based on configured volume.
+        // Note that since audioMultiplier_ is a Q5.11 fixed point number,
+        // the result pre-shift is a Q6.27 fixed point number. Shifting
+        // by 11 should cancel this out and result in the proper precision
+        // again.
+        dsps_mul_s16(tempAudioOut, audioMultiplier_, tempAudioOut, samplesToRead, 1, 1, 1, 11);
     
         auto packet = IcomPacket::CreateAudioPacket(
             audioSequenceNumber_++,
@@ -163,12 +345,11 @@ void AudioState::onAudioOutTimer_(DVTimer*)
 
 void AudioState::onRightChannelVolumeMessage_(DVTask* origin, storage::RightChannelVolumeMessage* message)
 {
-    float volInDb = 0.5 * message->volume;
+    short calcResult = FIXED_POINT_AMPLIFICATION_FACTORS[message->volume - MIN_AMPLIFICATION_DB];
 
-    float multiplier = std::exp(volInDb/20.0 * std::log(10.0));
     for (int index = 0; index < 160; index++)
     {
-        audioMultiplier_[index] = multiplier;
+        audioMultiplier_[index] = calcResult;
     }
 }